+++ ibis/expr/tests/test_format.py
@@ -250,7 +250,10 @@ CustomOp[CustomExpr]
     assert result == expected
 
 
-@pytest.mark.xfail(raises=AssertionError)
+@pytest.mark.xfail(
+    raises=AssertionError,
+    reason="PR 2283: The change for this test is reverted due to performance regression"
+)
 def test_same_column_multiple_aliases():
     table = ibis.table([('col', 'int64')], name='t')
     expr = table[table.col.name('alias1'), table.col.name('alias2')]
